// types.ts
export type ServiceType = 
  // Grupo Transporte/Conductores
  | 'Conductor elegido' | 'Conductor Familiar' | 'Trasporte para viajes' 
  | 'Trasporte de producto' | 'Transporte ocupantes al destino/domicilio' | 'translado documentos y mercancía'
  // Grupo Asistencia/Técnico
  | 'Abogado: Asistencia presencial' | 'Abogado: Asistencia telefónica' | 'Cerrajería autos' 
  | 'Carro taller por barada' | 'Carro taller instalación de batería' | 'Móvil presencial (CESVI)'
  // Grupo Grúas/Rescate
  | 'Deposito/custodia' | 'Grúa salida custodia' | 'Grua accidente' | 'Gra accedente pesados' 
  | 'Grua varada' | 'Grua varada pesados' | 'moto grua' | 'Rescate';